Expand object shorthand for imported names
In order to know when an expression is an object shorthand, we now walk the
tokens ahead of time and assign a context to each token so that we can
distinguish an identifier in an object literal vs an identifier in an array
literal, parameter list, etc. This seems to be enough information to make the
decisions we need to make.

This seems to work from a bunch of thinking, but certainly may still have flaws.
At the moment I assume that the codebase has semicolons; in order to loosen that
I'll need to implement ASI to distinguish `return \n {}` from `return {}`.