Add initial code for flow transform (#4)
This adds the flow transform as an option, adds a transformer (which does
nothing for now), and changes the import code to properly handle `type` in
various places. The behavior to copy from babel is that if the declaration is
`import type` or if the declaration has only `type` imports.

This does not yet try to handle the flow transform without the import transform.

The new transform is in the repl but marked as experimental since it's
incomplete. I probably won't publish until the transform is in better shape, but
the code is still publishable.