Remove flow function return types (#9)
We do this by looking at nearby tokens to see if it's an arrow function, normal
function, object literal function, or class body function. This also requires
carefully walking forward by one type, which is needed to distinguish an object
`{` from the start of the function body.

It may be better to compute this at parse time, or some similar system, but
hopefully I'll be able to get things working and then can iterate from there.