Allow class fields with names like "function"
We now isolate class field values as their own context, so if you're in a class
context you know that you won't be in a normal expression, and can treat
keywords like "function" as normal names. I also got rid of an assert that the
token is a name token, since in this case it isn't always one.