Label some identifiers with their roles (#71)
This makes it possible to distinguish an identifier access from declaration from
object shorthand, and simplifies a lot of fragile (and possibly wrong)
Sucrase-specific code. It may be best to move to a token-based analysis later,
but for now, it's cheap to attach the role as we pass over the identifier in
parsing.