Add an initial version of scopes emitted by the tokenizer
The hope here is to do minimal work at the lex/parse phase involving scopes, but
still enough that we can later identify which variables are shadowed via which
declarations. Rather than keeping a scope stack or anything like that, we just
emit a scope when we would have emitted a node, and have a flat list of scopes
that we pass into the normal Sucrase code. This can then be reconstructed in the
rare case that there actually is global shadowing in the file.

Also clean up some of the Options typings to avoid a use of `any`.