Propagate block or function scope declaration info to tokens
We now pass around a boolean for whether the assignment context is block-scoped,
and add that information on the tokens if so. That makes it possible to
determine the appropriate scope for a given variable.

This still misses bindings like function names and class declarations, so those
will come later.