Add a gulp plugin
This is the first non-core package, so it's a first pass at a monorepo. The
build and linting are all done at the top level and all are run in CI. I
intentionally named the folder "integrations" instead of "packages", since my
hope is that that will be the only use case for non-core packages. Hopefully,
those integration packages can be small and rarely-published, unlike the many
packages in the Babel monorepo.

Maybe later I'll explore Lerna and/or a Yarn workspace, but for now I'm keeping
it simple.