Add support for static class fields (#101)
This hasn't been thoroughly tested, but seems to work. At parse time, I save
whether the class is an expression or a statement, and use that to determine
whether we need to do the comma expression trick. When using a comma expression,
we need to generate a new variable name and also declare it at the top of the
file. For statement classes, we can just add some statements immediately after
the class definition, like we do for exports.