Fix the apollo build
There were some cases where Apollo was incorrectly relying on ES5 `var`
behavior, and another example where it was double-exporting the same thing, once
in a star export, which was causing problems due to a nuance around how star
exports are implemented. The star export thing may be nice to get 100% right in
the future, but it also I believe is a case of the code effectively relying on a
bug in TypeScript, so it doesn't seem especially worth supporting.