Fix import replacement for arrow function params (#131)
Arrow functions are parsed in a special way in Babylon where they are parsed as
comma expressions, then changed to a parameter list if it sees an arrow. The AST
is properly corrected, but the tokens aren't, so tokens weren't being marked as
declarations. This changes the strategy to backtrack if we see an arrow. This
requires another state.clone, but hopefully I can make that more efficient (or
unnecessary), and it should make it easier to get rid of node creation.