Get rid of toAssignable and some flow arrow function parsing (#133)
We shouldn't need to transform nodes anyway, and we now take an alternate
approach to ensuring that tokens are properly marked as declarations.

Also get rid of checkLVal.

There was also some recent Flow code that does a lot of try/catch operations
with nodes in a way that would be really hard to support with a simplified
parser. It was only added recently and isn't in the TypeScript parser, so it
seems likely safe to leave out, and I can find another approach if it ends up
being important.