Simplify the parser options and class hierarchy
Rather than dynamically building a class by chaining a subclass for each plugin,
we just have a fixed inheritance hierarchy for now, and in the future I'm hoping
to flatten it into plain functions rather than relying on inheritance.

Also get rid of options, which can now be replaced with just an array of plugin
names.