Remove token context code in favor of parser-directed tokenization
The parser is now responsible for declaring that a < should be treated as a JSX
start and that a / should be treated as a regex, which it knows because it knows
when it's at the start of an expression. That means that all of the context code
is no longer necessary, and we can remove those fields from the state.