Get an initial version of React working
This doesn't yet handle complex mutable export assignments, and instead just
changes to code to not use them in the one case where they're used, but it does
fix a few bugs involving flow and JSX, and it inserts development information
into JSX like the newer react preset does.

I also got rid of the react-display-name transform, in favor of the JSX
transform including it.