Handle some cases of complex export binding assignments
This follows the TypeScript strategy of replacing all variable usages with
`exports.foo`, which also ends up replacing assignments (including destructuring
assignments) as well. I tried to get the Babel strategy working, but properly
detecting assignees within destructure operations seems to be very difficult
with the current way that it's parsed. This change fixes the behavior for React,
and I believe the behavior now mimics TypeScript, so hopefully this should be a
good long-term solution.