Change token type to be a const enum
This seems to be only a slight performance improvement, but should also
hopefully make it easier to move to AssemblyScript. It makes debugging more
annoying because you just have an int for token types, so I'll have to see if
there's a nicer way to handle that, and maybe will go back to an object later,
but for now seems probably fine.