Switch from classes to plain function calls across the parser
This overhauls all parser code to use globals to track state, and replaces all
uses of method overriding with plain if statements. The idea is to reduce the
cost of function invocation, which is likely one of the biggest remaining costs
in the parser.

It looks like this gives about a 20% overall performance improvement and will
make it easier to move to AssemblyScript in the future.