Reimplement readWord as a code-generated decision tree
This seems to improve overall performance by about 15%. Rather than walking the
identifier, slicing a substring of the original input, and looking that
substring up, we now generate a decision tree in code that is able to match any
identifier (or contextual identifier) in a linear scan with no allocations.

This also changes the code generator to always run prettier on the result, so
that individual code generation functions don't need to worry about formatting.