Fix files accidentally included in final package
The example-repos directory contained a number .gitignore files that were
explicitly including files that we don't want to include, so we now clear all
example repos before publish.

Also, the implementation of `files` in package.json was including `dist`
directories in `integrations` advertently, so we now do the equivalent thing
in .npmignore with some fixes.