Move sucrase-babylon into the src directory (#242)
This simplifies some of the build step. I also got rid of the unused babelrc
file and renamed "parser" to "traverser" to avoid ambiguity with the overall
parser and to more clearly represent what the once-parser is now doing. This
also flattens the package structure so `dist` is analogous to `src`.

This isn't technically a breaking change, but anyone relying on the package file
structure could break from this, so I'll release it as a major version update
(with other breaking changes).