Fix crash on class fields that don't end in a semicolon (#271)
Fixes #235

The previous code always ate the next token after a class field, which was a
semicolon in all test cases so far. However, if the class field didn't end in a
semicolon, it would eat the first token of the next class body member, which
would confuse all downstream code and lead to errors.

This means we end up leaving semicolons around in class bodies where class
fields were, but this shouldn't cause problems, and it seems best to keep the
transforms simple.