Implement destructuring in export declarations (#305)
Fixes #297

The export transformation process now finds any declaration on the LHS,
excluding declarations within default values, and modifies them to use `exports`
syntax. In addition, for object destructures, we need to surround the entire
assignment with parens so that it's treated as a proper assignment. This
required some additional parser information: annotating the end index of an `=`
token in more cases, as well as identifying object shorthand declarations.