Properly handle function name inference in named exports
Fixes #307

Instead of writing `exports.f = () => {};`, we write
`const f = () => {}; exports.f = f;` and still use `exports.f` for all writes
and reads. This is a little odd in that the variable is never used, but it makes
sure the export works fine. In the future, we possibly could refine it to only
do a special case when the RHS is an unnamed function, but that's a bit harder
to detect.