Restrict TS try/catch logic to cases where backtracking is needed (#327)
Fixes #326

When porting some babel-parser changes, there was some code in TS parsing that
was unconditionally wrapping some parsing in a try/catch, but actually it only
cared about cases that started with `<`. By inverting the logic a little, we
avoid having to throw and catch an exception in the common case of parsing,
which was dramatically slowing down TypeScript parsing.