Fix syntax error line numbers, delay computation
We internally throw and catch syntax errors in a number of cases, and from some
profiling work, a lot of the time was spent generating the line number from the
source code offset, which the user will never see. Now, we only include the line
and column numbers when the exception is thrown from the top level of the
parser.

Also, the code to compute line numbers was completely broken anyway (everything
was showing up as on the first line), so this fixes it and adds a test.