Fix crash with TypeScript overloaded constructors (#340)
Fixes #334

Constructors without bodies were correctly being marked as type tokens, but the
class transformation code (to implement class fields and TS fields defined in
constructor params) was incorrectly viewing it as a real constructor. Now, when
checking if something matches the constructor keyword, we also make sure it's
not a type token.