Add microbenchmark, optimize isWhitespace
From some trial and error and printing out the generated machine code with
`node --print-opt-code`, it looks like switches aren't actually compiled all
that well and a lookup table performs a lot better, especially for the common
case of non-whitespace chars.

This optimization brings the microbenchmark running time from about 2 seconds to
about 1.2 seconds. With the function always returning true, it was 0.8 seconds,
so by that measure it's a 3x speedup.

I also changed benchmarks to compile on-the-fly since it doesn't seem to
actually affect perf much and it's nice to avoid the build step when iterating
on perf. I also simplified State construction a little.