Overhaul error management to avoid the need for try/catch backtracking (#347)
There were 13 places in the code where a try/catch was used for backtracking,
i.e. to try parsing one approach (e.g. JSX vs generic arrow function), catch
SyntaxError if there was a problem, then try the other approach. To some extent,
this parser complexity is needed due to the inherent difficulty in parsing
JS+JSX+TS (and Flow), particularly arrow functions and the many uses of `<`.
It also doesn't have a clear alternative, since any lookahead approach would
basically need to be a full parser to get all the details right.

try/catch is problematic not just for performance reasons but also because
exceptions aren't available in WebAssembly or Rust, both of which I want to
explore, so this change gets rid of all usages of try/catch.

Rather than changing every function in the parser to return an optional error
and propagate errors from its child calls (which would be a huge change and
likely hurt perf), this change swaps out `unexpected` so that rather than
throwing, it saves an error in `State` that can be accessed later. At that
point, code continues executing with the token state stuck in "EOF", and all
parsing operations are effectively no-ops. This is fragile because it means
that every function in the parser needs to be able to keep running in the
face of failure so that we can wind up the stack through normal function
completion. As part of this change, I audited all unbounded loops and made sure
they all break when in an error state. We still throw the saved error at the
top level, so the stack trace and error message is the same as before.