Fix playground hot reloading, simplify more webpack dev config (#385)
worker-plugin seems to cause an edit to its generated JS file on any edit to any
file, and it's never accepted as a hot module update, so webpack was refreshing
the page every time. To fix, I switched to use only-dev-server, so it still
shows a warning but doesn't refresh the page. Ideally it wouldn't be an edit or
would be accepted as a hot update, but this fixes things in practice.

I also got rid of some code that regularly clears the console, which makes it
easier to add console.logs to see what's happening within webpack.