Properly handle optional arrow params without type parameters
Fixes #387

I reworked `typedParseConditional` significantly from the Babel implementation
and didn't realize that I needed to preserve bare `?` tokens to denote the
parameter optional. Fortunately, the optional case will always be `?,` or `?)`
(an initializer isn't allowed when the param is optional), and neither of those
cases is valid as the start of a ternary, so we can distinguish optional param
from ternary with only one token of lookahead.