Switch codebase to use isolatedModules (#404)
We needed this anyway for Sucrase to work right, and it was just compiling const
enums as regular enums anyway, so this should have no effect on the resulting
code. This also means we no longer ship const enums in the public typedefs,
though ideally nobody was relying on those.

The main annoyance was that generating declarations is a little more annoying
because there are three flags to override from the CLI command.