Treat reserved words as invalid identifiers when handling enums
Fixes #392

There was already a code path for keys that couldn't be treated as regular
identifiers; with those, the variable declaration is omitted and we create a
special expression if necessary to add 1 to the previous value. Now, even for
regular identifiers, we see if the resulting identifier would be a JS reserve
word and treat it as a non-identifier if so. I had a few ideas on how to
reliably get a list of reserved words, but decided to just reference MDN since
there were enough edge cases with trying to modify the parser code.