Update README, including new swc perf comparison
I installed swc and hacked the benchmark to run it on the codebase used for the
comparison (the Benchling frontend codebase). There were about 100 places in the
code that crashed the swc parser, so I commented those out or otherwise modified
the code to get it working. I ran swc 10 times and took the fastest time, which
was 4.7x slower than Sucrase. One caveat is that swc was targeting ES3, since it
doesn't have an option to target latest JS.

Also mention the Broccoli integration, fixes #413.

Also fix a few other copy issues in the README.