Consolidate helper code into its own file
Now that we have four helpers and will soon be adding a fifth, it's nice to not
need a variable and lazy-getter for each one, so this puts them all in a file
that defines all helpers and coordinates those details.

This might be too dynamic to work in AssemblyScript, in which case I'd switch it
to use a codegen approach instead.