Elide type-only exports when the TS transform is enabled
Fixes #398

This ended up having a *lot* of cases and complexity. We actually care about the
known type declarations and the known value declarations.
* In ESM mode, only elide exports that are known to be a type and not known to
  be a value. This is true both for named export syntax and default export.
* In CJS mode, elide named exports that are either completely unknown or known
  to be a type. For default exports, it's the same logic as ESM: we only elide
  if we know it as a type and not as a value.

This required going through all declarations (at least in TS parsing mode) and
making sure we add the right identifier info to each of them, which in turn had
consequences on the react-hot-loader transform, so that had to be fixed.

One challenge is that default exports should be elided only when they export a
single identifier, which we can't know statically. Instead, we now use the
`rhsEndIndex` field on the `export` token to know when the export statement
actually ends, and use that to determine if the export is just an identifier.