Elide import = statements that are only used as a type (#441)
Fixes #422

In both CJS and ESM, we now look at the identifier in `import A =` and decide
whether to elide the entire statement in a similar way to normal TS import
elision. Since there are only two forms that these statements can have, I just
special cased the two when deleting statements.

Technically to implement this totally correctly, I'd need something like a graph
traversal: in an `import A = B.C;` statement, `B` is referenced as a value if
`A` is ever referenced as a value, so deciding if an import should be elided
requires following an arbitrary number of import statements. In practice, I
think this syntax is only ever used to pull in types (for values, you can just
use `const`, and the syntax is obscure anyway), so I treat it as a type
statement that gets elided.