Add runtime validation for options (#468)
Fixes #467

This uses the libraries ts-interface-builder and ts-interface-checker to codegen
a data structure of the options TS type and then check it on every `transform`
invocation.

I tried a few approaches here, including generating JSON schema and then
validating that, but that this seemed to be pretty smooth from a build system
standpoint.

I timed this on 1000 compilations of a large file and any difference was well
within the time variability, so it seems fine to run this type validation on
every `transform` invocation. It could potentially be moved out into a separate
routine if perf is ever an issue, though.