Update all dependencies to latest, update babel example (#474)
This runs `npx yarn-upgrade-all` and then regenerates the lockfile for each
project.

The babel example build was broken, so this updates it to latest. I had to make
a number of changes to get this to work, most notably getting rid of the few
usage of optional chaining syntax that have been added so far.

I didn't update webpack-object-rest-spread-plugin, since updating dependencies
caused a breakage and I probably want to just retire it soon enough.