First pass at implementation of optional chaining
The parser now associates each access operation with the start token for the
chain, and marks each chain containing an optional chain operation so that we
can wrap it in a function call.

This does not yet handle optional deletion, and there are a number of follow-up
tests that would be good to write, but the feature seems to be working with this
implementation.