Add initial test262 support to test optional chaining
For now, we only run two subdirectories and ignore various tests (sloppy mode,
tests expecting syntax error, and TCO tests). This reveals a few issues with the
current implementation, which will be addressed later. Since it's not a passing
build yet, I'll leave it out of CI for now.

Also drop the Travis build for Node 8 since one of the new dependencies requires
Node 10. I won't officially drop Node 8 support just yet, but will soon, and
it's at end-of-life now.