Fix nullish coalescing when the RHS is an object literal
Fixes #505

The RHS is meant to be transpiled into an expression arrow function, but if the
expression is an object literal, it gets interpreted as a block body. To fix, we
can unconditionally surround the arrow body in parens. This is a little ugly in
that it leaves an extra space, but avoiding that is probably more trouble than
it's worth.