Fix parsing of methods and fields named "declare"
Fixes #545

When originally porting https://github.com/babel/babel/pull/11146, there was a
mistake where we were using `eatContextual` when actually it was important to
use `tsParseModifier`, since `tsParseModifier` won't eat a name token if the
following token indicates that it must be the name of a method or field. This PR
fixes that by switching to `tsParseModifier` as Babel does.

It might be good to later refactor/simplify `tsParseModifier` if the keyword
specificity is just for error handling (and to maybe avoid the
snapshot/restore), but that can be a follow-up.