Add CI job to compare perf before and after and leave a comment
This should add some discipline to future perf improvements and help quantify
the negative perf impact of different changes when there is one. It runs the
Jest benchmark on the branch code and on the default branch code, alternating
between the two and doing five runs each and taking the median.

In order to properly report perf on PRs from forks, we save an artifact from the
main CI build and download it in a new `workflow_run` build which runs with
write access from the default branch code.

This article goes though a lot of the rationale and details:
https://securitylab.github.com/research/github-actions-preventing-pwn-requests