Add more benchmark scripts
This adds a jest-dev script to test the Jest dataset in better conditions to
measure the impact of code changes in Sucrase, plus a jest-diff script that is
similar and outputs JSON so that it can be used by an upcoming CI script to
measure performance before and after.

Also add some docs and clean up the package.json scripts.