Update immer dependency for website (#595)
CRA is still in the process of updating with the immer fix
( https://github.com/facebook/create-react-app/pull/10412 ), so this uses a
resolution override to use the newest immer. To reduce chance of compatibility
issues, I also updated `react-dev-utils` to latest and confirmed that dev and
prod builds both still work.