Add source map support to Jest plugin
Fixes #660

Previously, the Jest plugin would just return source code without attempting to
generate and return source maps. Even though the source map is always the
identity map, the lack of source map meant that WebStorm didn't allow debugger
breakpoints. To fix, we can use the same strategy as in `register.ts`: ask for a
source map, generate a base64 string of it, and append it to the end of the file
in a data URL.

One nuance here is that this adds a dependency on node because it uses the
`Buffer` class to convert to base64. I believe this is non-breaking because Jest
always runs in node anyway.

I was able to test this successfully in WebStorm and confirm that breakpoints
now work. Jest caches transform results, so I had to clear its cache for pick up
the change:
```
./node_modules/.bin/jest --clearCache
```